+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.2 - 2026-05-18
+
+### Fixed
+
+- Fixes separator placement and default picker ordering when MO2's left pane is
+  sorted by descending priority.
+
 ## 0.1.1 - 2026-05-14
 
 ### Fixed